// frontend/src/types/socket.types.ts
export enum SocketEvent {
  CONNECT = 'connect',
  DISCONNECT = 'disconnect',
  AVAILABLE_ROOMS = 'available_rooms',
  ROOM_CREATED = 'room_created',
  ROOM_JOINED = 'room_joined',
  ROOM_STATE_UPDATE = 'room_state_update',
  GAME_STARTED = 'game_started',
  GAME_STATE_UPDATE = 'game_state_update',
  PLAYER_GUESS_BROADCAST = 'player_guess_broadcast',
  GUESS_RESULT = 'guess_result',
  GAME_FINISHED = 'game_finished',
  ERROR = 'error',
  EXCEPTION = 'exception',
  CREATE_ROOM = 'create_room',
  JOIN_ROOM = 'join_room',
  SET_READY = 'set_ready',
  ADD_BOT = 'add_bot',
  REMOVE_BOT = 'remove_bot',
  LEAVE_ROOM = 'leave_room',
  GUESS_SUBMIT = 'guess_submit',
  GET_ROOMS = 'get_rooms',
}